+++ Testing/Source/MessageExchangeDefinition/Cxx/TestEcho.cxx
@@ -23,7 +23,7 @@
 int TestEcho(int , char *[])
 {
   std::string hostname = "www.dicomserver.co.uk";
-  int port = 11112;
+  uint16_t port = 11112;
   std::string callaetitle = "GDCM_ROCKS";
   std::string callingaetitle = "ACME1";
